+++ packages/wallet/src/initialization/instances/index.ts
@@ -1,1 +1,3 @@
-export { keyringController } from './keyring-controller';
+export { approvalController } from './approval-controller/approval-controller';
+export { keyringController } from './keyring-controller/keyring-controller';
+export { storageService } from './storage-service/storage-service';