+++ packages/wallet/src/types.ts
@@ -1,4 +1,3 @@
-import { KeyringControllerOptions } from '@metamask/keyring-controller';
 import type { Json } from '@metamask/utils';
 
 import type {
@@ -6,7 +5,9 @@ import type {
   DefaultEvents,
   RootMessenger,
 } from './initialization/defaults';
-import { GenericEncryptor } from './initialization/instances/keyring-controller';
+import type { ApprovalControllerInstanceOptions } from './initialization/instances/approval-controller/types';
+import type { KeyringControllerInstanceOptions } from './initialization/instances/keyring-controller/types';
+import type { StorageServiceInstanceOptions } from './initialization/instances/storage-service/types';
 import { InitializationConfiguration } from './initialization/types';
 
 export type WalletOptions = {
@@ -16,12 +17,11 @@ export type WalletOptions = {
     unknown,
     unknown
   >[];
-  instanceOptions?: InstanceSpecificOptions;
+  instanceOptions: InstanceSpecificOptions;
 };
 
 export type InstanceSpecificOptions = {
-  keyringController?: {
-    encryptor?: GenericEncryptor;
-    keyringBuilders?: KeyringControllerOptions['keyringBuilders'];
-  };
+  approvalController?: ApprovalControllerInstanceOptions;
+  keyringController?: KeyringControllerInstanceOptions;
+  storageService: StorageServiceInstanceOptions;
 };